Also have it "suggest" what to install
ya know, in the case that you dont have bluez and the rest of the gang installed.

!Minor rant!

Why the hell does bluez also install like, 70MB~ of fucking printer drivers? What the hell
is wrong with a sytstem that by default wants to install tons and tons of crap that only
~1% of people are going to use, even worse the end users are only going to use up to 2 types
of bloddy printer so why do I have to install 40 diffrent drivers to talk to diffrent printers
over bluetooth. argh.